refactoring step1
