+++ app/src/main/java/com/songlib/presentation/screens/selection/step1/Step1Screen.kt
@@ -8,11 +8,11 @@ import androidx.compose.runtime.*
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
-import com.songlib.data.models.Book
 import com.songlib.domain.entity.UiState
 import com.songlib.presentation.components.*
 import com.songlib.presentation.components.action.AppTopBar
 import com.songlib.presentation.navigation.Routes
+import com.songlib.presentation.screens.selection.step1.components.*
 import com.songlib.presentation.viewmodels.SelectionViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -82,42 +82,5 @@ fun Step1Screen(
                 )
             }
         }
-    )
-}
-
-@Composable
-fun Step1Fab(
-    viewModel: SelectionViewModel,
-    onSaveConfirmed: (List<Book>) -> Unit
-) {
-    var showConfirmDialog by remember { mutableStateOf(false) }
-    var showNoSelectionDialog by remember { mutableStateOf(false) }
-
-    if (showConfirmDialog) {
-        ConfirmSaveDialog(
-            onConfirm = {
-                onSaveConfirmed(viewModel.getSelectedBooks())
-                showConfirmDialog = false
-            },
-            onDismiss = { showConfirmDialog = false }
-        )
-    }
-
-    if (showNoSelectionDialog) {
-        NoSelectionDialog(
-            onDismiss = { showNoSelectionDialog = false }
-        )
-    }
-
-    ExtendedFloatingActionButton(
-        onClick = {
-            if (viewModel.getSelectedBooks().isNotEmpty()) {
-                showConfirmDialog = true
-            } else {
-                showNoSelectionDialog = true
-            }
-        },
-        icon = { Icon(Icons.Filled.Check, "Proceed") },
-        text = { Text(text = "Proceed") },
     )
 }
