+++ app/src/main/java/com/songlib/presentation/screens/selection/step1/components/Step1Fab.kt
@@ -1,0 +1,45 @@
+package com.songlib.presentation.screens.selection.step1.components
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import com.songlib.data.models.Book
+import com.songlib.presentation.viewmodels.SelectionViewModel
+
+@Composable
+fun Step1Fab(
+    viewModel: SelectionViewModel,
+    onSaveConfirmed: (List<Book>) -> Unit
+) {
+    var showConfirmDialog by remember { mutableStateOf(false) }
+    var showNoSelectionDialog by remember { mutableStateOf(false) }
+
+    if (showConfirmDialog) {
+        ConfirmSaveDialog(
+            onConfirm = {
+                onSaveConfirmed(viewModel.getSelectedBooks())
+                showConfirmDialog = false
+            },
+            onDismiss = { showConfirmDialog = false }
+        )
+    }
+
+    if (showNoSelectionDialog) {
+        NoSelectionDialog(
+            onDismiss = { showNoSelectionDialog = false }
+        )
+    }
+
+    ExtendedFloatingActionButton(
+        onClick = {
+            if (viewModel.getSelectedBooks().isNotEmpty()) {
+                showConfirmDialog = true
+            } else {
+                showNoSelectionDialog = true
+            }
+        },
+        icon = { Icon(Icons.Filled.Check, "Proceed") },
+        text = { Text(text = "Proceed") },
+    )
+}
